moved sales code to a module
