+++ src/lib/modules/sales/views/lead/index.jade
@@ -1,0 +1,49 @@
+extends ../../layout
+block head
+  include ../../datatables_css
+block content
+  .page-header
+    h1 Lead List
+  .container
+    table(id="table_leads").table.table-condensed.table-striped
+      thead
+  tr
+    th Name
+    th Company
+    th Email
+    th Phone
+    th Salespserson
+      tbody
+block footer
+  include ../../datatables_js
+  script(type='text/javascript').
+    $(document).ready(function() {
+      $('#table_leads').dataTable( {
+  "sDom": "<'row'<'span6'l><'span6'f>r>t<'row'<'span6'i><'span6'p>>",
+  "sPaginationType": "bootstrap",
+  "sAjaxSource": '/api/v1/sales/leads',
+  "iDisplayLength": 10,
+  "bStateSave": true,
+  "bDeferRender": true,
+  "bSortClasses": false,
+  "aoColumns": [
+    {
+      "mData": "full_name"
+    }, {
+      "mData": "company",
+      "mRender": function(company, type, row) {
+        return '<a href="/helpdesk/clients/clientid/' + row['clientid'] + '">' + company + '</a>';
+       }
+    }, {
+      "mData": "email"
+    }, {
+      "mData": "phone",
+    }, {
+      "mData": "salesperson_name",
+    }],
+  "aaSorting": [[1,'asc']],
+  "oLanguage": {
+    "sLengthMenu": "_MENU_ records per page"
+  }
+      });
+    });
