+++ src/lib/modules/sales/views/lead/new.jade
@@ -7,4 +7,4 @@ block content
   .container
     .row
       .col-md-6
-        include ../../partials/lead/leadform
+  include ../../partials/lead/leadform
