+++ src/routes/index.js
@@ -3,7 +3,7 @@ module.exports = function (app, config, authenticator, redisClient) {
 
   require('./auth')(app, config, authenticator, redisClient);
   require('./helpdesk')(app, config, authenticator, redisClient);
-  require('./sales')(app, config, authenticator, redisClient);
+  require(app.get('base_lib_dir') + '/modules/sales')(app, config, authenticator, redisClient);
   require('./monitoring')(app, config, authenticator, redisClient);
   require('./account')(app, config, authenticator, redisClient);
   require('./events')(app, config, authenticator, redisClient);
