--- src/views/sales/activity/index.jade
@@ -1,38 +1,0 @@
-extends ../../layout
-block head
-block content
-  .page-header
-    h1 Recent Sales Activity
-  .container
-    .col-md-12
-      .panel
-        .panel-heading Lead Activity
-        .panel-body
-          .row
-            table.table-condensed.table-hover.table-bordered.table-striped
-              thead
-                tr
-                  th Username
-                  th Client
-                  th Contact Method
-                  th Method Notes
-                  th Contact Disposition
-                  th Disposition Notes
-                  th Follow-Up
-                  th Follow-Up Notes
-                  th Follow-Up Scheduled
-              tbody
-                each activity in activities
-                  tr
-                    td #{activity.user.username}
-                    td
-                      a(href="/helpdesk/clients/clientid/#{activity.formData.clientID}") #{activity.formData.clientCompany}
-                    td #{activity.formData.contactMethod}
-                    td #{activity.formData.contactMethodNotes}
-                    td #{activity.formData.contactDisposition}
-                    td #{activity.formData.contactDispositionNotes}
-                    td #{activity.formData.contactFollowup}
-                    td #{activity.formData.contactFollowupNotes}
-                    td #{activity.formData.followupDate}
-          .row
-            a.btn.btn-primary(href="/sales/activity/new", type="button") Submit New Activity
