--- src/views/sales/index.jade
@@ -1,127 +1,0 @@
-extends ../layout
-block head
-  link(rel='stylesheet', href='/stylesheets/bootstrap-tile.css')
-block navbar
-block content
-  .container.col-xs-20.col-md-12
-    .row
-      .page-header
-        h1 Sales Team Dashboard
-    .row
-      .col-md-12
-        .panel.panel-default
-          .panel-heading Current Stats
-          .panel-body
-      .row
-        .col-md-6
-    .panel.panel-default
-      .panel-heading Opportunities Values
-      table.table.table-condensed.table-bordered
-        thead
-          tr
-      td &nbsp;
-      each stage,stage_id in metadata.stage_list
-        td #{stage}
-        tbody
-          tr
-      td Open
-      each stage,stage_id in metadata.stage_list
-        td #{stages[stage_id]['open']['sum']}
-          tr
-      td Won
-      each stage,stage_id in metadata.stage_list
-        td #{stages[stage_id]['won']['sum']}
-          tr
-      td Lost
-      each stage,stage_id in metadata.stage_list
-        td #{stages[stage_id]['lost']['sum']}
-      .row
-        .col.md-4
-    .panel.panel-default
-      .panel-heading Open Opportunities
-      .panel-body
-        each stage,stage_id in metadata.stage_list
-          div(id="stage#{stage_id}_open").tile.short.text-info
-      .tile-heading.short
-        span #{stage}
-        //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['open']['count']}
-        .tile-text.text-primary.short #{stages[stage_id]['open']['count']}
-        .col.md-4
-    .panel.panel-default
-      .panel-heading Won Opportunities
-      .panel-body
-        each stage,stage_id in metadata.stage_list
-          div(id="stage#{stage_id}_won").tile.short.text-info
-      .tile-heading.short
-        span #{stage}
-        //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['won']['count']}
-        .tile-text.text-primary.short #{stages[stage_id]['won']['count']}
-        .col.md-4
-    .panel.panel-default
-      .panel-heading Lost Opportunities
-      .panel-body
-        each stage,stage_id in metadata.stage_list
-          div(id="stage#{stage_id}_lost").tile.short.text-info
-      .tile-heading.short
-        span #{stage}
-        //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['lost']['count']}
-      .tile-text.text-primary.short #{stages[stage_id]['lost']['count']}
-    .row
-      .col-md-12
-        .panel.panel-default
-          .panel-heading Personal Status
-          .panel-body
-            .row
-              .col-md-6
-                .panel.panel-default
-                  .panel-heading Opportunities Values
-                  table.table.table-condensed.table-bordered
-                    thead
-                      tr
-                        td &nbsp;
-                        each stage,stage_id in metadata.stage_list
-                          td #{stage}
-                    tbody
-                      tr
-                        td Open
-                        each stage,stage_id in metadata.stage_list
-                          td #{stages[stage_id]['open']['sum']}
-                      tr
-                        td Won
-                        each stage,stage_id in metadata.stage_list
-                          td #{stages[stage_id]['won']['sum']}
-                      tr
-                        td Lost
-                        each stage,stage_id in metadata.stage_list
-                          td #{stages[stage_id]['lost']['sum']}
-            .row
-              .col.md-4
-                .panel.panel-default
-                  .panel-heading Open Opportunities
-                  .panel-body
-                    each stage,stage_id in metadata.stage_list
-                      div(id="stage#{stage_id}_open").tile.short.text-info
-                        .tile-heading.short
-                          span #{stage}
-                          //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['open']['count']}
-                          .tile-text.text-primary.short #{stages[stage_id]['open']['count']}
-              .col.md-4
-                .panel.panel-default
-                  .panel-heading Won Opportunities
-                  .panel-body
-                    each stage,stage_id in metadata.stage_list
-                      div(id="stage#{stage_id}_won").tile.short.text-info
-                        .tile-heading.short
-                          span #{stage}
-                          //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['won']['count']}
-                          .tile-text.text-primary.short #{stages[stage_id]['won']['count']}
-              .col.md-4
-                .panel.panel-default
-                  .panel-heading Lost Opportunities
-                  .panel-body
-                    each stage,stage_id in metadata.stage_list
-                      div(id="stage#{stage_id}_lost").tile.short.text-info
-                        .tile-heading.short
-                          span #{stage}
-                          //span.badge.primary.pull-right &nbsp; #{stages[stage_id]['lost']['count']}
-                        .tile-text.text-primary.short #{stages[stage_id]['lost']['count']}
